aaaaaa

+++ (1) basic/ft_strlcpy.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcpy(char *dest,const char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
 	unsigned int	i;
 	unsigned int	j;
@@ -18,7 +18,6 @@ unsigned int	ft_strlcpy(char *dest,const char *src, unsigned int size)
 	i = 0;
 	while (src[i])
 		i++;
-
 	if (size > 0)
 	{
 		j = 0;
